requirements.txt
langgraph 
langmem 
sentence_transformers 
pyyaml